requirements.txt
matplotlib==3.5.1
numpy==1.21.5
pandas==1.4.2
pandas_datareader==0.10.0
Pillow==9.1.1
plotly==5.8.0
prophet==1.0.1
streamlit==1.9.0
yfinance==0.1.70
